Add symbolic context
Add the ability to attach a symbolic context wrapper to the
planning pipeline. This involved an extremely long detour
through defining semantics for the composition and swapping
of wrapper metadata, and a refactoring of compile_plan.

It's a pretty cool system, although perhaps hard to debug,
and probably pretty hard to understand the inner workings...